Removed test file and moved logic
